refactor: UserProfileをServer/Client Componentへ分割 HOXBL-95
- 静的表示を'use client'を持たないUserProfileViewへ抽出
- ログアウト操作をLogoutButtonへ抽出しchildren経由で注入するCompositionパターンを適用
- console.log/console.errorをdebugLogロガーへ置換
- テストをUserProfileView.test.tsxとUserProfile.test.tsxに整理し重複を排除
